% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waitr.R
\name{waitr_delay}
\alias{waitr_delay}
\alias{waitr_until}
\alias{waitr_timestamp}
\alias{waitr_reset}
\title{Sleep for, or until a given time}
\usage{
waitr_delay(delay_ms)

waitr_until(time_ms)

waitr_timestamp()

waitr_reset()
}
\arguments{
\item{delay_ms}{Milliseconds to wait.}

\item{time_ms}{Clock time to wait for, in milliseconds.}
}
\value{
The time in milliseconds, relative to the last reset.
}
\description{
This package provides an internal clock measured in milliseconds since
the last reset, with functions to get the current timestamp, reset it to
zero, sleep for a number of milliseconds, or sleep until the relative time
is beyond a certain number of milliseconds. All functions return the relative
time in millseconds at the end of the function call.

This aim is to provide a lower overhead alternative to Sys.sleep, for
the purpose of simple animation at a reasonably consistent frame-rate,
regardless of platform and system.
}
